backend/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.14.0
django-cpf-cnpj==1.0.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
ipaddress==1.0.23
oauthlib==3.2.2
paramiko==3.1.0
phonenumbers==8.13.7
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pyodbc==4.0.39
python3-openid==3.2.0
pytz==2022.7.1
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.2
wcwidth==0.2.6
